backend/pyproject.toml
[project]
name = "api"
requires-python = ">=3.13"
version = "0.1.0"
dependencies = [
    "django>=5.1",
    "psycopg[binary]>=3.2",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "drf-spectacular>=0.28",
    "django-unfold>=0.43.0",
    "pillow>=10.0.0",
    # Stripe integration
    "stripe>=8.0.0",
    # Social Media OAuth and API dependencies
    "requests>=2.31.0",
    "python-facebook-api>=0.1.0",
    "tiktok-api>=0.1.0",
    "google-api-python-client>=2.108.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "linkedin-api>=2.0.0",
    "pinterest-api>=0.0.8",
    "instagram-private-api>=1.6.0",
    "cryptography>=41.0.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "django-celery-beat>=2.5.0",
    # OpenAI integration
    "openai>=1.54.0",
    "cloudinary>=1.44.1",
    "django-cloudinary-storage>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-factoryboy>=2.7.0",
]

[tool.ruff]
fix = true
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.pytest.ini_options]
filterwarnings = "ignore"
addopts = "--strict-config --strict-markers --ds=api.settings"
